[refactor] change mainActivity package
