+++ app/src/main/java/com/prac/githubrepo/login/LoginActivity.kt
@@ -3,18 +3,15 @@ package com.prac.githubrepo.login
 import android.content.Intent
 import android.net.Uri
 import android.os.Bundle
-import androidx.activity.enableEdgeToEdge
 import androidx.activity.viewModels
 import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
-import androidx.core.view.ViewCompat
-import androidx.core.view.WindowInsetsCompat
 import androidx.core.view.isVisible
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.lifecycleScope
 import androidx.lifecycle.repeatOnLifecycle
 import com.prac.githubrepo.BuildConfig
-import com.prac.githubrepo.MainActivity
+import com.prac.githubrepo.main.MainActivity
 import com.prac.githubrepo.R
 import com.prac.githubrepo.databinding.ActivityLoginBinding
 import dagger.hilt.android.AndroidEntryPoint
