+++ app/src/main/java/com/prac/githubrepo/main/MainActivity.kt
@@ -1,10 +1,8 @@
-package com.prac.githubrepo
+package com.prac.githubrepo.main
 
 import android.os.Bundle
-import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
-import androidx.core.view.ViewCompat
-import androidx.core.view.WindowInsetsCompat
+import com.prac.githubrepo.R
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
